hackathon/requirements.txt
google-adk>=1.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
protobuf~=5.29.4
psutil~=7.0.0
PyPDF2>=3.0.0
requests>=2.31.0